requirements.txt
paho-mqtt==1.5.0
PyYAML==5.3.1
